Refactor creature create sections into creature namespace
