+++ src/apps/library/create/creature/index.ts
@@ -1,2 +1,5 @@
 // src/apps/library/create/creature/index.ts
 export { CreateCreatureModal } from "./modal";
+export { mountCoreStatsSection } from "./section-core-stats";
+export { mountEntriesSection } from "./section-entries";
+export { mountSpellsKnownSection } from "./section-spells-known";
